fix sth + add gitignore

+++ src/main/java/com/chiaseyeuthuong/repository/SystemConfigRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface SystemConfigRepository extends JpaRepository<SystemConfig, Integer> {
+public interface SystemConfigRepository extends JpaRepository<SystemConfig, Long> {
 }
